# ARROW-8523: [C++] Optimize BitmapReader