Replacing bit offset with bit mask improves about 15% performance
with gcc-7.5. Arm64 servers have similar performance uplift.
clang-9 doesn't benefit from this change.

Below are arrow-bit-util-benchmark(BitmapReader/8192) results.
Comparing performance of "current code" -> "after this patch".

| cpu           | gcc-7.5    | clang-9    |
| ---           | -------    | -------    |
| Intel E5-2650 | 118 -> 146 | 117 -> 118 |
| Intel i7-4790 | 154 -> 191 | 155 -> 154 |
| AMD EPYC-7251 | 119 -> 133 | 122 -> 123 |