@cyb70289 2020-04-20
I forgot to add jira no in the first commit, modified later. Looks jira status is not synced with this PR.
Shall I abandon and push a new PR?
https://issues.apache.org/jira/browse/ARROW-8523