@cyb70289 2020-04-21
@pitrou @wesm 
Oops, I only checked case "BitmapReader" from benchmark [arrow-bit-util-benchmark](https://github.com/apache/arrow/blob/master/cpp/src/arrow/util/bit_util_benchmark.cc). Obviously it's not enough.

I compared all cases just now and see huge performance drop from below 4 tests:

Before this patch:
```bash
BenchmarkBitmapAnd/32768/1                 563496 ns       563260 ns         1243 bytes_per_second=55.4806M/s
BenchmarkBitmapAnd/131072/1               2219810 ns      2218984 ns          318 bytes_per_second=56.3321M/s
BenchmarkBitmapAnd/32768/2                 561738 ns       561467 ns         1265 bytes_per_second=55.6577M/s
BenchmarkBitmapAnd/131072/2               2246229 ns      2245119 ns          305 bytes_per_second=55.6763M/s
```

After this patch:
```bash
BenchmarkBitmapAnd/32768/1                1653467 ns      1652680 ns          422 bytes_per_second=18.9087M/s
BenchmarkBitmapAnd/131072/1               6665501 ns      6661561 ns          105 bytes_per_second=18.7644M/s
BenchmarkBitmapAnd/32768/2                1670793 ns      1670246 ns          423 bytes_per_second=18.7098M/s
BenchmarkBitmapAnd/131072/2               6702369 ns      6698957 ns          103 bytes_per_second=18.6596M/s
```

Before reverting this patch, I would like to understand why it happens.

BTW: we definitely need continuous benchmark tools to detect these things early.