@cyb70289 2020-04-21
Opened a jira card https://issues.apache.org/jira/browse/ARROW-8537