@pitrou 2020-04-21
To be honest, `BitmapAnd` should probably be rewritten using `Bitmap::VisitWords`.
But we can revert anyway if we fear regressions may appear in other workloads.