@wesm 2020-04-21
> BTW: we definitely need continuous benchmark tools to detect these things early.

Agreed. Hopefully some progress can be made on this in 2020 since the prior discussion in 2019 didn't go anywhere.