@wesm 2020-04-21
In the meantime, when we have microperformance patches like these it would be a good practice in the future to make sure that performance results are reproduced in the codebase's benchmark executables rather than on an ad hoc basis