@cyb70289 2020-04-21
@wesm, actually I did use codebase's benchmark executable. The problem is I only focused on one case that's directly related to this change. But ignored other cases that look not relevant, and finally found impacted.
Guess I can write a simple script to automate benchmark result checking as a temporary solution.