@wesm 2020-04-21
FWIW, we have some benchmark diffing code already written in 

https://github.com/apache/arrow/blob/master/dev/archery/archery/benchmark

I'm not sure where this is documented / how to use it to check the output of a single executable 

cc @fsaintjacques