@fsaintjacques 2020-04-21
See either `archery benchmark diff --help` or the [benchmark](https://arrow.apache.org/docs/developers/benchmarks.html) section of the documentation. Archery can compare the same binary with different scenarios:
- Commit
- Toolchain
- Compile flags

All it needs is two cmake build directory and it'll compare the benchmark binaries from there.